+++ cookbooks/homebrew/attributes/default.rb
@@ -1,1 +1,1 @@
-default["homebrew"]["user"]          = "jclyons"
+default["homebrew"]["user"]          = "administrator"